+++ bookclub/urls.py
@@ -93,6 +93,21 @@ urlpatterns = [
         name="logout",
     ),
     path(
+        "accounts/password_change/",
+        auth_views.PasswordChangeView.as_view(
+            template_name="bookclub/password_change_form.html",
+            success_url="/accounts/password_change/done/",
+        ),
+        name="password_change",
+    ),
+    path(
+        "accounts/password_change/done/",
+        auth_views.PasswordChangeDoneView.as_view(
+            template_name="bookclub/password_change_done.html"
+        ),
+        name="password_change_done",
+    ),
+    path(
         "register/<uuid:invite_code>/",
         register_with_invite,
         name="register_with_invite",